More example cleans
