+++ examples/block/00_block_no_payload.rs
@@ -24,6 +24,8 @@ async fn main() -> Result<()> {
     // Create and send the block.
     let block = client.block().finish().await?;
 
+    println!("{:#?}", block);
+
     println!(
         "Block with no payload sent: {}/block/{}",
         std::env::var("EXPLORER_URL").unwrap(),
