+++ examples/block/04_block_tagged_data.rs
@@ -1,0 +1,52 @@
+// Copyright 2021 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+//! This example sends a block with a tagged data payload.
+//! Run: `cargo run --example block_tagged_data --release -- [NODE URL]`.
+
+use iota_client::{block::payload::Payload, Client, Result};
+
+#[tokio::main]
+async fn main() -> Result<()> {
+    // Take the node URL from command line argument or use one from env as default.
+    let node_url = std::env::args().nth(1).unwrap_or_else(|| {
+        // This example uses dotenv, which is not safe for use in production.
+        dotenv::dotenv().ok();
+        std::env::var("NODE_URL").unwrap()
+    });
+
+    // Create a client with that node.
+    let client = Client::builder()
+        .with_node(&node_url)?
+        .with_node_sync_disabled()
+        .finish()?;
+
+    // Create and send the block with tag and data.
+    let block = client
+        .block()
+        .with_tag("Hello".as_bytes().to_vec())
+        .with_data("Tangle".as_bytes().to_vec())
+        .finish()
+        .await?;
+
+    println!("{:#?}\n", block);
+
+    if let Some(Payload::TaggedData(payload)) = block.payload() {
+        println!(
+            "Tag: {}",
+            String::from_utf8(payload.tag().to_vec()).expect("found invalid UTF-8")
+        );
+        println!(
+            "Data: {}",
+            String::from_utf8(payload.data().to_vec()).expect("found invalid UTF-8")
+        );
+    }
+
+    println!(
+        "\nBlock with tag and data sent: {}/block/{}",
+        std::env::var("EXPLORER_URL").unwrap(),
+        block.id()
+    );
+
+    Ok(())
+}
