+++ examples/block/custom_inputs.rs
@@ -55,6 +55,8 @@ async fn main() -> Result<()> {
         .finish()
         .await?;
 
+    println!("{:#?}", block);
+
     println!(
         "Transaction sent: {}/block/{}",
         std::env::var("EXPLORER_URL").unwrap(),
