+++ examples/block/output.rs
@@ -47,6 +47,8 @@ async fn main() -> Result<()> {
         .finish()
         .await?;
 
+    println!("{:#?}", block);
+
     println!("Transaction sent: {node_url}/api/core/v2/blocks/{}", block.id());
     println!("Block metadata: {node_url}/api/core/v2/blocks/{}/metadata", block.id());
