+++ examples/output/recursive_alias.rs
@@ -141,18 +141,20 @@ async fn main() -> Result<()> {
     //////////////////////////////////
     // create third transaction with the third alias output updated
     //////////////////////////////////
-    let outputs = vec![alias_output_builder
-        .clone()
-        .with_alias_id(alias_id_2)
-        .with_state_index(1)
-        .with_state_metadata(vec![3, 2, 1])
-        .replace_unlock_condition(UnlockCondition::StateControllerAddress(
-            StateControllerAddressUnlockCondition::new(alias_1_address),
-        ))?
-        .replace_unlock_condition(UnlockCondition::GovernorAddress(GovernorAddressUnlockCondition::new(
-            alias_1_address,
-        )))?
-        .finish_output()?];
+    let outputs = vec![
+        alias_output_builder
+            .clone()
+            .with_alias_id(alias_id_2)
+            .with_state_index(1)
+            .with_state_metadata(vec![3, 2, 1])
+            .replace_unlock_condition(UnlockCondition::StateControllerAddress(
+                StateControllerAddressUnlockCondition::new(alias_1_address),
+            ))?
+            .replace_unlock_condition(UnlockCondition::GovernorAddress(GovernorAddressUnlockCondition::new(
+                alias_1_address,
+            )))?
+            .finish_output()?,
+    ];
 
     let block_3 = client
         .block()
@@ -169,17 +171,19 @@ async fn main() -> Result<()> {
     //////////////////////////////////
     // create fourth transaction with the third alias output updated again
     //////////////////////////////////
-    let outputs = vec![alias_output_builder
-        .with_alias_id(alias_id_2)
-        .with_state_index(2)
-        .with_state_metadata(vec![2, 1, 3])
-        .replace_unlock_condition(UnlockCondition::StateControllerAddress(
-            StateControllerAddressUnlockCondition::new(alias_1_address),
-        ))?
-        .replace_unlock_condition(UnlockCondition::GovernorAddress(GovernorAddressUnlockCondition::new(
-            alias_1_address,
-        )))?
-        .finish_output()?];
+    let outputs = vec![
+        alias_output_builder
+            .with_alias_id(alias_id_2)
+            .with_state_index(2)
+            .with_state_metadata(vec![2, 1, 3])
+            .replace_unlock_condition(UnlockCondition::StateControllerAddress(
+                StateControllerAddressUnlockCondition::new(alias_1_address),
+            ))?
+            .replace_unlock_condition(UnlockCondition::GovernorAddress(GovernorAddressUnlockCondition::new(
+                alias_1_address,
+            )))?
+            .finish_output()?,
+    ];
 
     let block_3 = client
         .block()
